+++ example/a.ts
@@ -1,14 +1,10 @@
 import { Elysia, t } from '../src'
 
-new Elysia()
-	.ws('/ws', {
-		open(ws) {
-			const { query } = ws.data
-			console.log(query)
-		},
-		body: t.String({ minLength: 1 }),
-		error({ error }) {
-			console.log(error)
-		}
-	})
-	.listen(3000)
+const parserPlugin = new Elysia().onParse({ as: 'scoped' }, () => {})
+const deletePlugin = new Elysia().delete('/delete', () => ({
+	message: 'Resource deleted!'
+}))
+
+const app = new Elysia().use(parserPlugin).use(deletePlugin).listen(3000)
+
+console.log(app.routes[0].compile().toString())